* Large refactors:
** New curvature equation
The curvature equation is now based on hermite polynomials (generated via
s-power series).  Hopefully this will be the last time the curvature
equation changes.

** Multires
Support for multiresolution detail editing was adding, but is disabled.
I originally meant to use vector bsplines for this, but I may end up
using clothoid splines for it too.

On to some UI improvements!